Add retry option (#110)
* Add retry option
