+++ src/services/spiService/index.ts
@@ -19,7 +19,7 @@ import {
   updateTxFlow,
   updateTxMessage,
 } from '../../redux/reducers/TerminalSlice/terminalsSlice';
-import { getLocalStorage, setLocalStorage } from '../../utils/common/spi/common';
+import { getLocalStorage, setLocalStorage, getTxFlow } from '../../utils/common/spi/common';
 import SpiEventTarget from '../../utils/common/spi/eventTarget';
 import { ITerminal, ITerminals } from '../interfaces';
 
@@ -287,7 +287,7 @@ class SpiService {
           settings: null, // not available during pair terminal stage
           status: spiClient?._currentStatus,
           terminalStatus: spiClient?.CurrentFlow,
-          txFlow: spiClient?.CurrentTxFlowState,
+          txFlow: getTxFlow(spiClient?.CurrentTxFlowState),
           txMessage: null, // not available during pair terminal stage
         };
 
@@ -387,53 +387,10 @@ class SpiService {
             })
           );
 
-        const txFlowDetails = {
-          posRefId: detail?.PosRefId,
-          id: detail?.Id,
-          type: detail?.Type,
-          displayMessage: detail?.DisplayMessage,
-          amountCents: detail?.AmountCents,
-          awaitingSignatureCheck: detail?.AwaitingSignatureCheck,
-          finished: detail?.Finished,
-          success: detail?.Success,
-          response: {
-            data: {
-              rrn: detail?.Response?.Data?.rrn,
-              schemeAppName: detail?.Response?.Data?.scheme_app_name,
-              schemeName: detail?.Response?.Data?.scheme_name,
-              merchantReceipt: detail?.Response?.Data?.merchant_receipt,
-              transactionType: detail?.Response?.Data?.transaction_Type,
-              hostResponseText: detail?.Response?.Data?.host_response_text,
-            },
-          },
-          signatureRequiredMessage: detail?.SignatureRequiredMessage,
-          request: {
-            id: detail?.Request.Id,
-            eventName: detail?.Request?.EventName,
-            data: {
-              posRefId: detail?.Request?.Data?.pos_ref_id,
-              purchaseAmount: detail?.Request?.Data?.purchase_amount,
-              tipAmount: detail?.Request?.Data?.tip_amount,
-              cashAmount: detail?.Request?.Data?.cash_amount,
-              promptForCashout: detail?.Request?.Data?.prompt_for_cashout,
-              surchargeAmount: detail?.Request?.Data?.surcharge_amount,
-              promptForCustomerCopy: false,
-              printForSignatureRequiredTransactions: false,
-              printMerchantCopy: false,
-              customerReceiptHeader: '',
-              customerReceiptFooter: '',
-              merchantReceiptHeader: '',
-              merchantReceiptFooter: '',
-            },
-            posId: '',
-            decryptedJson: '',
-          },
-        };
-
         this.dispatchAction(
           updateTxFlow({
             id: instanceId,
-            txFlow: txFlowDetails,
+            txFlow: getTxFlow(detail),
           })
         );
       });
